Split hw-app-btc, refactor in async/await, progress events
