+++ packages/hw-app-btc/package.json
@@ -27,6 +27,7 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@ledgerhq/hw-transport": "^5.2.0",
+    "bip32-path": "^0.4.2",
     "create-hash": "^1.1.3"
   },
   "devDependencies": {
