+++ script/watch.sh
@@ -5,5 +5,8 @@ set -e
 cd ../..
 PATH=$(yarn bin):$PATH
 cd -
-babel --watch --source-maps -d lib src &
-flow-copy-source -w -v src lib
+export NODE_ENV=production
+BABEL_ENV=cjs babel --source-maps --watch -d lib src --config-file ../../babel.config.js &
+flow-copy-source -v src lib  -w &
+BABEL_ENV=es babel --source-maps --watch -d lib-es src --config-file ../../babel.config.js &
+flow-copy-source -v src lib-es -w
